% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation_wrapper.R
\name{create_quag_ph_spez}
\alias{create_quag_ph_spez}
\title{Create a document from template quag_ph_spez}
\usage{
create_quag_ph_spez(psDocuName, psPkgPath = ".", psDocuSubdir = "vignettes",
  pDocuHasOwnSubdir = TRUE, pbOverwrite = FALSE, pbEdit = TRUE,
  plReplace = NULL)
}
\arguments{
\item{psDocuName}{name of the new document}

\item{psPkgPath}{path where package is located under which document should be created}

\item{psDocuSubdir}{subdirectory in which document should be saved to}

\item{pbOverwrite}{flag whether existing files are overwritten}

\item{pbEdit}{directly open newly created document}

\item{plReplace}{list with replacement key-values}

\item{pbDocuHasOwnSubdir}{should document be stored in separate subdir}
}
\description{
This function is a wrapper to create a document that produces
pdf output and that uses the template "quag_ph_spez"
}
\details{
Per default the new document is created in a separate subdirectory,
is not overwritten if it already exists, is open to be edited
directly and does not pass on any information for pattern
replacement.
}
\examples{
\dontrun{
rqudocfmt::create_quag_ph_spez(psDocuName = "my_doc")
}
}
