% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/creation_wrapper.R
\name{create_qupps_de}
\alias{create_qupps_de}
\title{Create a document from template qualitas_project_ph_spez_de}
\usage{
create_qupps_de(psDocuName, psPkgPath, psDocuSubdir = "vignettes",
  pDocuHasOwnSubdir = TRUE, pbOverwrite = FALSE, pbEdit = TRUE,
  plReplace = NULL)
}
\arguments{
\item{psDocuName}{name of the new document}

\item{psPkgPath}{path where package is located under which document should be created}

\item{psDocuSubdir}{subdirectory in which document should be saved to}

\item{pbOverwrite}{flag whether existing files are overwritten}

\item{pbEdit}{directly open newly created document}

\item{plReplace}{list with replacement key-values}

\item{pbDocuHasOwnSubdir}{should document be stored in separate subdir}
}
\description{
This function is a wrapper to create a document that produces
pdf output and that uses the template "qualitas_project_ph_spez_de"
}
\details{
Per default the new document is created in a separate subdirectory,
is not overwritten if it already exists, is open to be edited
directly and does not pass on any information for pattern
replacement.
}

